Move move generator to Board
